refactor(GameMode): :recycle: game mode types
moved file to game mode folder/package.
cleaned game mode types.
