+++ duelinvaders/src/main/java/org/enstabretagne/Game/GameLauncher.java
@@ -27,8 +27,8 @@ import org.enstabretagne.Component.AlienComponent;
 import org.enstabretagne.Component.SpaceInvadersFactory;
 import org.enstabretagne.Game.GameModes.ClassicGameMode;
 import org.enstabretagne.Game.GameModes.GameMode;
+import org.enstabretagne.Game.GameModes.GameModeTypes;
 import org.enstabretagne.Utils.EntityType;
-import org.enstabretagne.Utils.GameModeTypes;
 import org.enstabretagne.Utils.GameVariableNames;
 import org.enstabretagne.Utils.Settings;
 import org.enstabretagne.Utils.assetNames;
