+++ duelinvaders/src/main/java/org/enstabretagne/Game/GameModes/GameMode.java
@@ -1,7 +1,5 @@
 package org.enstabretagne.Game.GameModes;
 
-import org.enstabretagne.Utils.GameModeTypes;
-
 import com.almasb.fxgl.input.Input;
 
 public interface GameMode {
